# backend-lecture-code/day17-microservices/microservices/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Day 17 Store — Microservices (via the API Gateway)
  version: 1.0.0
  description: |
    The same store as the monolith, split into eight services behind one gateway. The public
    paths are identical to the monolith's; only the port changed (4100 → 8080).

    **Log in first:** call `POST /users/login` with `{"email": "alice@shop.com"}`, copy the
    `token`, then click **Authorize** and paste it. The gateway checks the token once and passes
    the caller's id to the service in an `x-user-id` header. Services never see the token.

    Any routed request can also fail at the gateway itself: **502** when the service is down,
    **504** when it takes longer than 10 seconds.

    The **internal** section at the bottom lists service-to-service endpoints. The gateway has no
    route for them; they're documented with each service's own address for reference.

    Seed data to know: the 85" TV (`SKU-2001`, $5,999) is over the $5,000 limit, so paying for
    it is always declined. The Coffee Mug (`SKU-3001`) has only 1 in stock.
servers:
  - url: http://localhost:8080
    description: API gateway
...
paths:
...
  /users:
    post:
      tags: [users]
      summary: Sign up
      security: [] # public, no token needed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NewUser" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /products:
    get:
      tags: [catalog]
      summary: List products
      security: [] # public, no token needed
      description: |
        Round-robins across every catalog instance in `CATALOG_URLS`. Start a second instance
        and watch the gateway log alternate between them.
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string, example: electronics }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
        "502": { $ref: "#/components/responses/ServiceDown" }
    post:
      tags: [catalog]
      summary: Add a product
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Product" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /inventory:
    get:
      tags: [inventory]
      summary: Stock for every SKU
      security: [] # public, no token needed
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Stock" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /cart:
    get:
      tags: [cart]
      summary: My cart
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
    delete:
      tags: [cart]
      summary: Empty my cart
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: Emptied }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /orders:
    post:
      tags: [orders]
      summary: Check out my cart
      description: |
        The orders service calls five other services, in order:

        1. **cart** and **users** (in parallel): the items and the customer's email
        2. **catalog**: current prices, copied into the order as a snapshot
        3. **inventory**: reserve stock (committed on inventory's side immediately)
        4. saves the order in its own database as `pending_payment`
        5. **payments**: charge the card. If this fails or times out (3s), orders **releases the
           reservation itself** and marks the order `payment_failed`
        6. **cart**: empty it (best effort)
        7. **notifications**: fire-and-forget. If it's down, the order still succeeds.
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402":
          description: Card declined (any total over $5,000). The stock was released and the order is saved as `payment_failed`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "409":
          description: Not enough stock
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "502": { $ref: "#/components/responses/ServiceDown" }
        "503":
          description: A service that checkout depends on is unreachable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
        "504":
          description: A service that checkout depends on timed out (e.g. payments after 3s)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UpstreamError" }
    get:
      tags: [orders]
      summary: My orders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
  /notifications:
    get:
      tags: [notifications]
      summary: Emails sent to me
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Email" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/ServiceDown" }
  /recommendations:
    get:
      tags: [recommendations, demo]
      summary: Three recommended products (slow on purpose, in Python)
      security: [] # public, no token needed
      description: |
        Burns CPU for `ms` milliseconds, the same as the monolith's version. This time it runs in
        its own process, so `GET /products` keeps answering instantly while this is busy.
      parameters:
        - name: ms
          in: query
          required: false
          schema: { type: integer, default: 5000, maximum: 30000 }
      responses:
        "200":
          description: OK. If catalog is down, `recommendations` is empty and `note` explains why.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Recommendations" }
        "502": { $ref: "#/components/responses/ServiceDown" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid, or expired token (checked at the gateway), or unknown email on login
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: Already exists, or not enough stock
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceDown:
      description: The service behind this route is unreachable (from the gateway itself)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/UpstreamError" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, example: "unknown sku NOPE" }
    UpstreamError:
      type: object
      properties:
        error: { type: string, example: "payments timed out after 3000ms" }
        service: { type: string, example: payments }
...
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, example: alice@shop.com }
        name: { type: string, example: Alice }
    NewUser:
      type: object
      required: [email, name]
      properties:
        email: { type: string, example: dave@shop.com }
        name: { type: string, example: Dave }
...
    Product:
      type: object
      required: [sku, name, price, category]
      properties:
        sku: { type: string, example: SKU-2001 }
        name: { type: string, example: 85" 8K TV }
        price: { type: number, example: 5999 }
        category: { type: string, example: electronics }
        attributes:
          type: object
          description: Different for every kind of product (stored in MongoDB, so no fixed schema).
          additionalProperties: true
          example: { resolution: 8K, screenInches: 85 }
    Stock:
      type: object
      properties:
        sku: { type: string, example: SKU-1001 }
        onHand: { type: integer, example: 25 }
...
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              sku: { type: string, example: SKU-1001 }
              quantity: { type: integer, example: 2 }
    Order:
      type: object
      description: Includes a snapshot of the customer and items as they were at checkout.
      properties:
        id: { type: integer, example: 1 }
        userId: { type: integer, example: 1 }
        status:
          type: string
          enum: [pending_payment, placed, payment_failed]
          example: placed
        total: { type: number, example: 194.48 }
        customerEmail: { type: string, example: alice@shop.com }
        items:
          type: array
          items:
            type: object
            properties:
              sku: { type: string, example: SKU-1001 }
              name: { type: string, example: Running Shoes }
              quantity: { type: integer, example: 2 }
              unitPrice: { type: number, example: 89.99 }
        reservationId: { type: integer, example: 1 }
        paymentId: { type: integer, nullable: true, example: 1 }
        createdAt: { type: string, format: date-time }
...
    Email:
      type: object
      properties:
        id: { type: integer, example: 1 }
        userId: { type: integer, example: 1 }
        to: { type: string, example: alice@shop.com }
        subject: { type: string, example: "Order #1 confirmed" }
        body: { type: string, example: "Total $194.48" }
        sentAt: { type: string, format: date-time }
    Recommendations:
      type: object
      properties:
        tookMs: { type: integer, example: 5000 }
        servedBy: { type: string, example: recommendations (Python / FastAPI) }
        note: { type: string, example: catalog unreachable }
        recommendations:
          type: array
          items:
            type: object
            properties:
              sku: { type: string }
              name: { type: string }
              price: { type: number }